requirements.txt
Django==5.2.5
psycopg2-binary==2.9.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
whitenoise==6.7.0
gunicorn==22.0.0
uvicorn==0.30.6

